# docs/swagger.yaml
components:
  schemas:
...
    api.followStatusResponse:
      properties:
        follower_count:
          description: 该用户粉丝数
          minimum: 0
          type: integer
        viewer_following:
          description: 是否已关注
          type: boolean
      required:
      - follower_count
      - viewer_following
      type: object
    api.followUserResponse:
      properties:
        avatar_url:
          description: 头像地址
          type: string
        bio:
          description: 个人简介
          type: string
        id:
          description: 用户 ID
          type: string
        username:
          description: 用户名
          type: string
        viewer_following:
          description: 当前用户是否已关注
          type: boolean
      required:
      - id
      - viewer_following
      type: object
...
    api.likeStatusResponse:
      properties:
        like_count:
          description: 点赞数量
          minimum: 0
          type: integer
        viewer_liked:
          description: 是否已点赞
          type: boolean
      required:
      - like_count
      - viewer_liked
      type: object
...
    api.pageResponse-api_followUserResponse:
      properties:
        items:
          items:
            $ref: '#/components/schemas/api.followUserResponse'
          type: array
          uniqueItems: false
        page:
          minimum: 1
          type: integer
        page_size:
          minimum: 1
          type: integer
        total:
          minimum: 0
          type: integer
      required:
      - items
      - page
      - page_size
      - total
      type: object
...
    render.Response-api_followStatusResponse:
      properties:
        code:
          type: integer
        data:
          $ref: '#/components/schemas/api.followStatusResponse'
        msg:
          type: string
      type: object
...
    render.Response-api_likeStatusResponse:
      properties:
        code:
          type: integer
        data:
          $ref: '#/components/schemas/api.likeStatusResponse'
        msg:
          type: string
      type: object
...
    render.Response-api_pageResponse-api_followUserResponse:
      properties:
        code:
          type: integer
        data:
          $ref: '#/components/schemas/api.pageResponse-api_followUserResponse'
        msg:
          type: string
      type: object
...
    render.errorResponse:
      properties:
        code:
          type: integer
        msg:
          type: string
      type: object
  securitySchemes:
    BearerAuth:
      scheme: bearer
      type: http
...
info:
  description: 小蓝书后端接口文档
  title: 小蓝书 API
  version: "1.0"
openapi: 3.1.0
paths:
...
  /comments/{comment_id}/like:
    delete:
      parameters:
      - description: 评论 ID
        in: path
        name: comment_id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.Response-api_likeStatusResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.errorResponse'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.errorResponse'
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: 取消点赞评论
      tags:
      - likes
    put:
      parameters:
      - description: 评论 ID
        in: path
        name: comment_id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.Response-api_likeStatusResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.errorResponse'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.errorResponse'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.errorResponse'
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: 点赞评论
      tags:
      - likes
...
  /posts/{post_id}/like:
    delete:
      parameters:
      - description: 帖子 ID
        in: path
        name: post_id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.Response-api_likeStatusResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.errorResponse'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.errorResponse'
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: 取消点赞帖子
      tags:
      - likes
    put:
      parameters:
      - description: 帖子 ID
        in: path
        name: post_id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.Response-api_likeStatusResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.errorResponse'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.errorResponse'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.errorResponse'
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: 点赞帖子
      tags:
      - likes
...
  /users/{user_id}/follow:
    delete:
      parameters:
      - description: 目标用户 ID
        in: path
        name: user_id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.Response-api_followStatusResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.errorResponse'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.errorResponse'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.errorResponse'
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: 取消关注
      tags:
      - follows
    put:
      parameters:
      - description: 目标用户 ID
        in: path
        name: user_id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.Response-api_followStatusResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.errorResponse'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.errorResponse'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.errorResponse'
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: 关注用户
      tags:
      - follows
  /users/{user_id}/followers:
    get:
      parameters:
      - description: 用户 ID
        in: path
        name: user_id
        required: true
        schema:
          type: string
      - description: 页码
        in: query
        name: page
        schema:
          default: 1
          type: integer
      - description: 每页数量
        in: query
        name: page_size
        schema:
          default: 20
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.Response-api_pageResponse-api_followUserResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.errorResponse'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.errorResponse'
          description: Internal Server Error
      summary: 获取粉丝列表
      tags:
      - follows
  /users/{user_id}/following:
    get:
      parameters:
      - description: 用户 ID
        in: path
        name: user_id
        required: true
        schema:
          type: string
      - description: 页码
        in: query
        name: page
        schema:
          default: 1
          type: integer
      - description: 每页数量
        in: query
        name: page_size
        schema:
          default: 20
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.Response-api_pageResponse-api_followUserResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.errorResponse'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.errorResponse'
          description: Internal Server Error
      summary: 获取关注列表
      tags:
      - follows
...
servers:
- description: 本地开发环境
  url: http://localhost:8080/api/v1